+++ src/DateTime/DateTime.py
@@ -1249,12 +1249,10 @@ class DateTime:
         """
         if t is None:
             t = 0
-        if isinstance(t, float):
+        if isinstance(t, (float, int)):
             return self._micros > long(t * 1000000)
-        try:
+        else:
             return self._micros > t._micros
-        except AttributeError:
-            return self._micros > t
 
     __gt__ = greaterThan
 
@@ -1272,12 +1270,10 @@ class DateTime:
         """
         if t is None:
             t = 0
-        if isinstance(t, float):
+        if isinstance(t, (float, int)):
             return self._micros >= long(t * 1000000)
-        try:
+        else:
             return self._micros >= t._micros
-        except AttributeError:
-            return self._micros >= t
 
     __ge__ = greaterThanEqualTo
 
@@ -1294,12 +1290,10 @@ class DateTime:
         """
         if t is None:
             t = 0
-        if isinstance(t, float):
+        if isinstance(t, (float, int)):
             return self._micros == long(t * 1000000)
-        try:
+        else:
             return self._micros == t._micros
-        except AttributeError:
-            return self._micros == t
 
     def notEqualTo(self, t):
         """Compare this DateTime object to another DateTime object
@@ -1341,12 +1335,10 @@ class DateTime:
         """
         if t is None:
             t = 0
-        if isinstance(t, float):
+        if isinstance(t, (float, int)):
             return self._micros < long(t * 1000000)
-        try:
+        else:
             return self._micros < t._micros
-        except AttributeError:
-            return self._micros < t
 
     __lt__ = lessThan
 
@@ -1363,12 +1355,10 @@ class DateTime:
         """
         if t is None:
             t = 0
-        if isinstance(t, float):
+        if isinstance(t, (float, int)):
             return self._micros <= long(t * 1000000)
-        try:
+        else:
             return self._micros <= t._micros
-        except AttributeError:
-            return self._micros <= t
 
     __le__ = lessThanEqualTo